refactor: moving bp´s to routes
refactor: moving bp´s to routes
